requirements.txt
anyio==3.5.0
APScheduler==3.9.1
asttokens==2.0.5
backcall==0.2.0
backports.entry-points-selectable==1.1.1
cachetools==5.0.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.7
click==8.1.2
colorama==0.4.4
decorator==5.1.1
distlib==0.3.3
executing==0.8.3
ffmpeg==1.4
filelock==3.3.2
googletrans==3.0.0
h11==0.12.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.14.7
httpx==0.22.0
hyperframe==5.2.0
idna==2.10
ipython==8.3.0
jedi==0.18.1
libretranslatepy==2.1.1
lxml==4.8.0
matplotlib-inline==0.1.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.4.0
prompt-toolkit==3.0.29
pure-eval==0.2.2
pyaes==1.6.1
pyasn1==0.4.8
Pygments==2.12.0
python-telegram-bot==20.0a0
pytube==12.0.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
stack-data==0.2.0
telebot==0.0.4
Telethon==1.24.0
tornado==6.1
traitlets==5.1.1
translate==3.6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.7
virtualenv==20.10.0
wcwidth==0.2.5
